Bonus_apps/requirements.txt
langchain
langchain_google_genai
Pillow
reportlab
pdf2image